Add a finall output color pass and fix window resizing/camera changes in the editor

+++ GloryEngine/Editor/GloryEditor/GameWindow.h
@@ -17,7 +17,7 @@ namespace Glory::Editor
         void View();
 
     private:
-        size_t m_CurrentOutputCameraIndex;
+        int m_CurrentOutputCameraIndex;
         int m_SelectedRenderTextureIndex;
     };
 }
